Finish the move from _ to not _
