+++ webviz_subsurface/containers/_parameter_distribution.py
@@ -13,7 +13,7 @@ from webviz_config import WebvizContainerABC
 from webviz_config.common_cache import CACHE
 from webviz_config.webviz_store import webvizstore
 
-from .._datainput import load_parameters
+from .._datainput.fmu_input import load_parameters
 
 
 class ParameterDistribution(WebvizContainerABC):
